fix proto

+++ app/chat/internal/service/chat.go
@@ -336,7 +336,7 @@ func (l *Chat) createConversation(ctx context.Context, owner, target string, con
 		pipe.Expire(ctx, key, time.Duration(constant.ConvKeepDays*24)*time.Hour)
 
 		conv := &common.Conversation{
-			Type:   convType,
+			Type:   int32(convType),
 			Target: target,
 		}
 
@@ -438,7 +438,7 @@ func (l *Chat) Recall(ctx context.Context, req *chat.RecallReq, rsp *chat.Recall
 	}
 	b, _ := json.Marshal(m)
 	reqL := chat.SendReq{
-		ConvType: v.ConvType,
+		ConvType: int32(v.ConvType),
 		MsgType:  constant.MsgRecall,
 		Sender:   v.Sender,
 		Target:   v.Target,
